Refactor: Move stateless services to utils/ (#53)
- Move sessionService.js → utils/sessionExtractor.js
- Move userService.js → utils/userExtractor.js
- Move validationService.js → utils/requestValidator.js
- Move fileProcessorService.js → utils/fileProcessor.js
- Update all imports
- Document convention in AGENTS.md: services = classes, utils = functions
- Add file uploads feature to README.md
