+++ src/utils/sessionExtractor.js
